requirements.txt
python-telegram-bot[job-queue]==21.11.1
feedparser==6.0.11
httpx==0.28.1
python-dotenv==1.1.0
